Finished

+++ MyRMI/src/rmipacket/RemoteObjectRef.java
@@ -1,5 +1,6 @@
-
 package rmipacket;
+
+
 
 import java.io.ByteArrayInputStream;
 import java.io.ByteArrayOutputStream;
@@ -13,48 +14,47 @@ import java.net.UnknownHostException;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
-
 public class RemoteObjectRef implements Serializable {
-    
+
     private InetAddress ip;
     private int port;
-    private long time;
+    private float time;
     private int objectNumber;
     public RemoteObjectInterface remoteInterface;
-    
-    public RemoteObjectRef(){
-        
+
+    public RemoteObjectRef() {
+
     }
 
-    public RemoteObjectRef(String host, int port, long time, int objectNumber, Object remoteInterface) throws UnknownHostException{
+    public RemoteObjectRef(String host, int port, float time, int objectNumber, Object remoteInterface) throws UnknownHostException {
         this.ip = InetAddress.getByName(host);
         this.port = port;
         this.time = time;
         this.objectNumber = objectNumber;
         this.remoteInterface = new RemoteObjectInterface(remoteInterface);
     }
-    
-    public Method getMethod(Object obj, int methodNumber){
+
+    public Method getMethod(Object obj, int methodNumber) {
         return remoteInterface.getMethod(obj, methodNumber);
     }
-    
-    public int getObjNumber(){
+
+    public int getObjNumber() {
         return objectNumber;
     }
-    
-    public RemoteObjectInterface getRemoteInterface(){
+
+    public RemoteObjectInterface getRemoteInterface() {
         return remoteInterface;
     }
-    
-    public InetAddress getAddress() throws UnknownHostException{
+
+    public InetAddress getAddress() throws UnknownHostException {
         return ip;
     }
-    
-    public int getPort(){
+
+    public int getPort() {
         return port;
     }
-    
-    public byte[] toByte(){
+
+    public byte[] toByte() {
         ObjectOutputStream os = null;
         try {
             ByteArrayOutputStream byteStream = new ByteArrayOutputStream();
@@ -65,7 +65,7 @@ public class RemoteObjectRef implements Serializable {
             return byteStream.toByteArray();
         } catch (IOException ex) {
             Logger.getLogger(RemoteObjectRef.class.getName()).log(Level.SEVERE, null, ex);
-        } finally{
+        } finally {
             try {
                 os.close();
             } catch (IOException ex) {
@@ -74,14 +74,14 @@ public class RemoteObjectRef implements Serializable {
         }
         return null;
     }
-    
+
     public RemoteObjectRef toRemoteObjectRef(byte[] b) {
         ObjectInputStream os = null;
         try {
             ByteArrayInputStream byteStream = new ByteArrayInputStream(b);
             os = new ObjectInputStream(byteStream);
             return (RemoteObjectRef) os.readObject();
-            
+
         } catch (IOException ex) {
             Logger.getLogger(RemoteObjectRef.class.getName()).log(Level.SEVERE, null, ex);
         } catch (ClassNotFoundException ex) {
@@ -94,8 +94,7 @@ public class RemoteObjectRef implements Serializable {
             }
         }
         return null;
-        
+
     }
-    
-    
+
 }
